+++ src/ach-example.c
@@ -73,7 +73,7 @@ ach_channel_t chan_feedback;
 double now() {
     struct timespec t;
     clock_gettime( CLOCK_MONOTONIC, &t );
-    return t.tv_sec + t.tv_nsec / 1e9 ;
+    return (double)(t.tv_sec) + (double)t.tv_nsec / 1e9 ;
 }
 
 /* simple integrator, x = dt * dx */
